introduce-pool-stats-observer-pre-move: Move PoolStats dataclass to scheduler_components.pool_stats_observer
Pure dataclass move pre-move for ``introduce-pool-stats-observer``.

Cut the ``PoolStats`` dataclass body byte-identical from
``scheduler_runtime_checker_mixin.py`` and paste it into the new
module ``scheduler_components/pool_stats_observer.py``. Rewire the
mixin to re-import ``PoolStats`` from the new module so existing
references continue to resolve. The test
``test/registered/unit/managers/test_scheduler_pause_generation.py``
also imports ``PoolStats`` directly; its import path is updated in this
commit.

Refactor chain ID: introduce-pool-stats-observer-pre-move
